bug(applications): the meaning of 'app' and 'application' changed
update the code to reflect this.
Whilst most of this is cosmetic, the commands to run `jx delete app` are now out of date and failing (because the command used to be `jx delete app` (meaning delete an application) but has now changed to be `jx delete application` so as not to conflict with the entirely different `jx delete app` (meaning delete a platform addon) command.