fix: Let tide merge apps and UI test PRs.
Ideally, I'd like to do this with the upgrade boot test too, but I'll
come back to that - there's no `--auto-merge` flag there, so it would
need to have a `/approve` added, which may not work quite right with
the bot user. I'll loop back on that.

fixes https://github.com/jenkins-x/jx/issues/6850

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>