fix: Use /approve for merging PRs
This will require that the `BDD_APPROVER_USERNAME` and
`BDD_APPROVER_ACCESS_TOKEN` environment variables be set. Otherwise,
the UI and apps tests will just rely on the `--auto-merge` flag.

The upgrade boot test will continue to merge manually, but is now
using `GitProvider` functions instead of the github client library,
just to decrease complexity/need for multiple ways of loading the GH
token, etc.

The approver needs to be a different user than the bot user, and will
be added as a collaborator on the repo for any PR it tries to merge.

fixes https://github.com/jenkins-x/jx/issues/6861

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>